ED-19279 Tweak: Kit Screenshots
