+++ app/modules/kit-library/module.php
@@ -4,7 +4,6 @@ namespace Elementor\App\Modules\KitLibrary;
 use Elementor\App\Modules\KitLibrary\Data\Repository;
 use Elementor\Core\Admin\Menu\Admin_Menu_Manager;
 use Elementor\Core\Admin\Menu\Main as MainMenu;
-use Elementor\Modules\CloudLibrary\Render_Mode_Preview_Base;
 use Elementor\Plugin;
 use Elementor\TemplateLibrary\Source_Local;
 use Elementor\Core\Base\Module as BaseModule;
@@ -120,19 +119,42 @@ class Module extends BaseModule {
 			add_action( 'elementor/kit-library/generate-screenshot', [ $this, 'handle_kit_screenshot' ], 10, 1 );
 			add_action( 'template_redirect', [ $this, 'handle_kit_screenshot_generation' ] );
 		}
-		add_action( 'elementor/kit-library/generate-screenshot', [ $this, 'handle_kit_screenshot' ], 10, 1 );
-		add_action( 'template_redirect', [ $this, 'handle_kit_screenshot_generation' ] );
 	}
 
 	public function handle_kit_screenshot() {
-		$config = [
-			'home_url' => home_url(),
-			'kit_id' => uniqid(),
-		];
+		$suffix = ( Utils::is_script_debug() || Utils::is_elementor_tests() ) ? '' : '.min';
 
 		show_admin_bar( false );
 
-		Render_Mode_Preview_Base::enqueue_scripts_for_screenshot( [ 'elementor-common', 'elementor-common-modules' ] );
+		wp_enqueue_script(
+			'dom-to-image',
+			ELEMENTOR_ASSETS_URL . "/lib/dom-to-image/js/dom-to-image{$suffix}.js",
+			[],
+			'2.6.0',
+			true
+		);
+
+		wp_enqueue_script(
+			'html2canvas',
+			ELEMENTOR_ASSETS_URL . "/lib/html2canvas/js/html2canvas{$suffix}.js",
+			[],
+			'1.4.1',
+			true
+		);
+
+		wp_enqueue_script(
+			'cloud-library-screenshot',
+			ELEMENTOR_ASSETS_URL . "/js/cloud-library-screenshot{$suffix}.js",
+			[ 'dom-to-image', 'html2canvas', 'elementor-common', 'elementor-common-modules' ],
+			ELEMENTOR_VERSION,
+			true
+		);
+
+		$config = [
+			'home_url' => home_url(),
+			'kit_id' => uniqid(),
+			'selector' => 'body',
+		];
 
 		wp_add_inline_script( 'cloud-library-screenshot', 'var ElementorScreenshotConfig = ' . wp_json_encode( $config ) . ';' );
 	}
